+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
   * raised minimum required PHP version to 7.0.0
   * introduced scalar type hints and strict type checking
+  * calling `stubbles\peer\http\HttpUri::fromString()` with an empty string now throws `stubbles\peer\MalformedUri` instead of returning `null`
 
 
 7.1.0 (2016-06-08)